Updates & revisions for closer parity to iOS
