+++ .gitattributes
@@ -1,0 +1,1 @@
+android/**/snapshots/**/*.png filter=lfs diff=lfs merge=lfs -text
