+++ android/build.gradle
@@ -1,10 +1,11 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
-    id 'com.android.application' version '8.4.0' apply false
-    id 'com.android.library' version '8.4.0' apply false
-    id 'org.jetbrains.kotlin.android' version '1.9.22' apply false
+    id 'com.android.application' version '8.4.2' apply false
+    id 'com.android.library' version '8.4.2' apply false
+    id 'org.jetbrains.kotlin.android' version '1.9.24' apply false
     id 'com.github.willir.rust.cargo-ndk-android' version '0.3.4' apply false
     id 'com.ncorti.ktfmt.gradle' version '0.17.0' apply false
+    id 'app.cash.paparazzi' version '1.3.4' apply false
     id 'maven-publish'
 }
