+++ android/composeui/src/androidTest/java/com/stadiamaps/ferrostar/composeui/LocalizedDistanceFormatterDistanceMeasurementSystemOverrideTest.kt
@@ -1,6 +1,8 @@
 package com.stadiamaps.ferrostar.composeui
 
 import android.icu.util.ULocale
+import com.stadiamaps.ferrostar.composeui.formatting.DistanceMeasurementSystem
+import com.stadiamaps.ferrostar.composeui.formatting.LocalizedDistanceFormatter
 import org.junit.Assert
 import org.junit.Test
