+++ android/core/build.gradle
@@ -41,19 +41,19 @@ android {
 }
 
 dependencies {
-    implementation(platform("com.squareup.okhttp3:okhttp-bom:4.10.0"))
+    implementation(platform("com.squareup.okhttp3:okhttp-bom:4.11.0"))
 
     // For as long as we support API 25; once we can raise support to 26, all is fine
     coreLibraryDesugaring 'com.android.tools:desugar_jdk_libs:2.0.4'
 
-    implementation 'androidx.core:core-ktx:1.12.0'
-    implementation 'androidx.appcompat:appcompat:1.6.1'
+    implementation 'androidx.core:core-ktx:1.13.1'
+    implementation 'androidx.appcompat:appcompat:1.7.0'
     implementation 'com.squareup.okhttp3:okhttp'
     implementation 'com.squareup.moshi:moshi:1.15.1'
-    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.7.3'
-    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.7.0'
+    implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-core:1.8.1'
+    implementation 'androidx.lifecycle:lifecycle-viewmodel-ktx:2.8.3'
 
-    implementation "net.java.dev.jna:jna:5.12.0@aar"
+    implementation "net.java.dev.jna:jna:5.12.1@aar"
 
     testImplementation 'junit:junit:4.13.2'
 
@@ -62,8 +62,8 @@ dependencies {
     androidTestImplementation 'com.github.gmazzo:okhttp-mock:1.5.0'
     androidTestImplementation 'org.jetbrains.kotlinx:kotlinx-coroutines-test:1.7.3'
 
-    androidTestImplementation 'androidx.test.ext:junit:1.1.5'
-    androidTestImplementation 'androidx.test.espresso:espresso-core:3.5.1'
+    androidTestImplementation 'androidx.test.ext:junit:1.2.1'
+    androidTestImplementation 'androidx.test.espresso:espresso-core:3.6.1'
 }
 
 cargoNdk {
